IsoMix

Documentation for IsoMix.

IsoMix.Component

IsoMix.Component1

IsoMix.Component2

IsoMix.Component3

IsoMix.Constant

IsoMix.Data

IsoMix.Data1

IsoMix.Data2

IsoMix.Data3

IsoMix.Datum

IsoMix.Fraction

IsoMix.Fraction2

IsoMix.Fraction3

IsoMix.IsoMixType

IsoMix.Model

IsoMix.Model1

IsoMix.Model2

IsoMix.Model3

IsoMix.Norm

IsoMix.Prior

IsoMix.Prior2

IsoMix.Prior3

IsoMix.System

IsoMix.System2

IsoMix.System3

IsoMix.Unf

IsoMix.logNorm

IsoMix.countcomponents

IsoMix.fractions

IsoMix.mix

IsoMix.mix!

IsoMix.Component — Type

Component <: IsoMixType

Abstract supertype for Component_ instances, where the _ indicates dimensionality. Represents the composition of a component or endmember within a natural system.

see also: Component1, Component2, Component3

Component(x, cx) -> Component1

Component(x, cx, y, cy) -> Component2

Component(x, cx, y, cy, z, cz) -> Component3

The constructor function accepts values as a list of arguments or as keyword assignments, e.g.: Component(x=1, cx=2)

IsoMix.Component1 — Type

Component1 <: Component

1-dimensional Component instance.

Fields

x	Isotopic composition of x

cx	Concentration of x

see also: Component

IsoMix.Component2 — Type

Component2 <: Component

2-dimensional Component instance.

Fields

x	Isotopic composition of x

cx	Concentration of x

y	Isotopic composition of y

cy	Concentration of y

see also: Component

IsoMix.Component3 — Type

Component3 <: Component

3-dimensional Component instance.

Fields

x	Isotopic composition of x

cx	Concentration of x

y	Isotopic composition of y

cy	Concentration of y

z	Isotopic composition of z

cz	Concentration of z

see also: Component

IsoMix.Constant — Type

Constant(x) <: Datum

A Datum instance with a discrete value x.

IsoMix.Data — Type

Data <: IsoMixType

Abstract supertype for Data_ instances, where the _ indicates dimensionality.

see also: Data1, Data2, Data3

IsoMix.Data1 — Type

Data1 <: Data

1-dimensional Data instance.

Fields

x	Isotopic composition of x

cx	Concentration of x

see also: Datum

IsoMix.Data2 — Type

Data2 <: Data

2-dimensional Data instance.

Fields

x	Isotopic composition of x

cx	Concentration of x

y	Isotopic composition of y

cy	Concentration of y

see also: Datum

IsoMix.Data3 — Type

Data3 <: Data

3-dimensional Data instance.

Fields

x	Isotopic composition of x

cx	Concentration of x

y	Isotopic composition of y

cy	Concentration of y

z	Isotopic composition of z

cz	Concentration of z

see also: Datum

IsoMix.Datum — Type

Datum <: IsoMixType

Abstract supertype for Datum instances.

see also: Constant, Unf, Norm, logNorm

IsoMix.Fraction — Type

Fraction <: IsoMixType

Abstract supertype for Fraction_ instances that contain vectors of fractional combinations of _ components/endmembers.

see also: Fraction2, Fraction3

Fraction(Amin, Amax; n=101)

Returns a Fraction2 with n linearly spaced fractional combinations of components A and B that each sum to 1, given minimum and maximum A fractions of Amin and Amax, respectively. Sincen gives the linear fractions of each component, this constructor returns n-length vectors of each component. Both Amin and Amax must fall within [0,1].

Example

julia> Fraction(0,1,n=3)
Fraction2([0.0, 0.5, 1.0], [1.0, 0.5, 0.0], 3)

Fraction(Amin, Amax, Bmin, Bmax; n=101)

Returns a Fraction3 with linearly spaced fractional combinations of components A, B, and C that each sum to 1, given respective minima and maxima of A (Amin and Amax) and B (Bmin and Bmax). Since n gives the linear fractions of each component, this constructor returns a n(n÷2+1)-length vector for each component. Amin,Amax, Bmin, and Bmax must fall within [0,1].

Example

julia> f=Fraction(0,1,0,1,n=3)
Fraction3([0.0, 0.0, 0.0, 0.5, 0.5, 1.0], [0.0, 0.5, 1.0, 0.0, 0.5, 0.0], [1.0, 0.5, 0.0, 0.5, 0.0, 0.0], 6)

Fraction(D, n=101)

Given a provided number of components D (2 or 3), returns a corresponding Fraction instance with component concentrations of 0 to 1.

Example

julia> Fraction(2,n=3)
Fraction2([0.0, 0.5, 1.0], [1.0, 0.5, 0.0], 3)

IsoMix.Fraction2 — Type

Fraction2 <: Fraction

2-dimensional Fraction instance.

Fields

A	Endmember/component A

B	Endmember/component B

see also: Fraction

IsoMix.Fraction3 — Type

Fraction3 <: Fraction

3-dimensional Fraction instance.

Fields

A	Endmember/component A

B	Endmember/component B

C	Endmember/component C

see also: Fraction

IsoMix.IsoMixType — Type

IsoMixType

Supertype containing all custom types and structs in the IsoMix package:

IsoMixType

├─ Component

│ ├─ Component1

│ ├─ Component2

│ └─ Component3

├─ Data

│ ├─ Data1

│ ├─ Data2

│ └─ Data3

├─ Datum

│ ├─ Constant

│ ├─ Norm

│ ├─ Unf

│ └─ logNorm

├─ Fraction

│ ├─ Fraction2

│ └─ Fraction3

├─ Model

│ ├─ Model1

│ ├─ Model2

│ └─ Model3

├─ Prior

│ ├─ Prior2

│ └─ Prior3

└─ System

├─ System2

└─ System3

Example

julia> using AbstractTrees,IsoMix; AbstractTrees.children(d) = subtypes(d); print_tree(IsoMixType)

IsoMix.Model — Type

Model <: IsoMixType

Abstract supertype for Model_ instances, where the _ indicates dimensionality.

see also: Model1, Model2, Model3

Model(f::Fraction, s::System)

Generate a Model instance

IsoMix.Model1 — Type

Model1 <: Model

1-dimensional Model instance: a single species tracking isotopic composition and concentration.

Fields

x	Isotopic composition of x

cx	Concentration of x

see also: Model

IsoMix.Model2 — Type

Model2 <: Model

2-dimensional Model instance: two species tracking isotopic composition and concentration.

Fields

x	Isotopic composition of x

cx	Concentration of x

y	Isotopic composition of y

cy	Concentration of y

see also: Model

IsoMix.Model3 — Type

Model3 <: Model

3-dimensional Model instance: three species tracking isotopic composition and concentration.

Fields

x	Isotopic composition of x

cx	Concentration of x

y	Isotopic composition of y

cy	Concentration of y

z	Isotopic composition of z

cz	Concentration of z

see also: Model

IsoMix.Norm — Type

Norm(m, s) <: Datum

Normally distributed Datum with mean m and standard deviation s.

IsoMix.Prior — Type

Prior <: IsoMixType

Abstract supertype for Prior_ instances, where the _ indicates dimensionality. Represents a suite of data that provide a prior estimate for the endmember compositions of a natural system.

see also: Prior2, Prior3

Prior(A::Data, B::Data)

Returns a Prior2

Prior(A::Data, B::Data, C::Data)

Returns a Prior3

Note: each Data must be of the same subtype (i.e. consistent number of components within each endmember of a system).

IsoMix.Prior2 — Type

Prior2 <: Prior

2-dimensional Prior instance.

Fields

A	Endmember/component A

B	Endmember/component B

see also: Prior

IsoMix.Prior3 — Type

Prior3 <: Prior

3-dimensional Prior instance.

Fields

A	Endmember/component A

B	Endmember/component B

C	Endmember/component C

see also: Prior

IsoMix.System — Type

Prior <: IsoMixType

Abstract supertype for System_ instances, where the _ indicates dimensionality. Represents a model system composition charaterized its Component fields.

see also: Prior2, Prior3

The constructor function accepts values as a list of arguments:

System(A::Component, B::Component)

Returns a System2

System(A::Component, B::Component, C::Component)

Returns a System3

... or as keyword assignments, e.g.: System(A=..., B=...)

IsoMix.System2 — Type

System2 <: System

2-dimensional System instance.

Fields

A	Endmember/component A

B	Endmember/component B

see also: System

IsoMix.System3 — Type

System3 <: System

3-dimensional System instance.

Fields

A	Endmember/component A

B	Endmember/component B

C	Endmember/component C

see also: System

IsoMix.Unf — Type

Unf(a, b) <: Datum

Uniformly distributed Datum with minimum a and maximum b, inclusive. I.e., for any $x$ in $Unf(a, b)$, $x \in [a,b]$.

IsoMix.logNorm — Type

logNorm(lm, ls) <: Datum

Log-normally distributed Datum with log-mean lm and log-space standard deviation ls.

IsoMix.countcomponents — Method

IsoMix.countcomponents(c::Component)

Count the number of elemental dimensions (1-3) of c.

Example

julia> IsoMix.countcomponents(Component(1,2,3,4,5,6))
3

IsoMix.fractions — Method

fractions(s<:System2{Component1}, x)

Calculate the fractional contributions of each the two Components in s, given a composition of x.

fractions(s<:System3{Component2}, x, y)

Calculate the fractional contributions of each the three Components in s, given compositions of x and y.

IsoMix.mix! — Method

mix!(m <: Model, s <: System, f <: Fraction)

In-place version of mix that overwrites fields in m.

see also: System, Fraction, Model

IsoMix.mix — Method

mix(s<:System, f<:Fraction)

Calculate mixing between components of system s given fractional mixtures f, returned in a Model instance.

see also: mix, System, Fraction, Model

Input configurations:

mix(s::System2{Component1}, f::Fraction2) -> Model1

Two-endmember, 1 species.

mix(s::System2{Component2}, f::Fraction2) -> Model2

Two-endmember, 2 species.

mix(s::System2{Component3}, f::Fraction2) -> Model3

Two-endmember, 3 species.

mix(s::System3{Component2}, f::Fraction3) -> Model2

Three-endmember, 2 species.

mix(s::System3{Component3}, f::Fraction3) -> Model3

Three-endmember, 3 species.