IsoMix

Documentation for IsoMix.

IsoMix.Constant

IsoMix.Data

IsoMix.Datum

IsoMix.DistributionDraws

IsoMix.EmDraw

IsoMix.EmDraw1

IsoMix.EmDraw2

IsoMix.EmDraw3

IsoMix.Endmember

IsoMix.Endmember1

IsoMix.Endmember2

IsoMix.Endmember3

IsoMix.Fraction

IsoMix.Fraction2

IsoMix.Fraction3

IsoMix.IsoMixType

IsoMix.Measurements

IsoMix.Measurements1

IsoMix.Measurements2

IsoMix.Measurements3

IsoMix.Model

IsoMix.Model1

IsoMix.Model2

IsoMix.Model3

IsoMix.Norm

IsoMix.Prior

IsoMix.Prior2

IsoMix.Prior3

IsoMix.SysDraw

IsoMix.SysDraw2

IsoMix.SysDraw3

IsoMix.Unconstrained

IsoMix.Unf

IsoMix.logNorm

IsoMix.countcomponents

IsoMix.extractcomponents

IsoMix.extractfields

IsoMix.extractsystem

IsoMix.fractions

IsoMix.initialguess

IsoMix.initialjump

IsoMix.jump

IsoMix.loglikelihood

IsoMix.mix

IsoMix.mix!

IsoMix.nanll

IsoMix.reassigncomponents

IsoMix.stopwatch

IsoMix.top3

IsoMix.update

IsoMix.Constant — Type

Constant(x) <: Datum

A Datum instance with a discrete value x.

IsoMix.Data — Type

Data <: IsoMixType

Abstract supertype for types that contain measured data, represented by Datum instances. Prior instances contain 2-3 Endmember instances, and both Endmember orMeasurementsinstances containDatum` instances.

see also: Datum, Endmember, Prior, Measurements

IsoMix.Datum — Type

Datum <: Data

Abstract supertype for Datum instances.

see also: Constant, Unf, Norm, logNorm, Unconstrained

IsoMix.DistributionDraws — Type

DistributionDraws <: IsoMixType

Abstract supertype for types that contain values drawn from Data distributions of Endmember and Prior instances. SysDraw instances contain 2-3 EmDraw instances, which contain discrete draws from a corresponding Endmember instance.

see also: EmDraw, SysDraw

IsoMix.EmDraw — Type

EmDraw <: DistributionDraws

Abstract supertype for EmDraw_ instances, where the _ indicates dimensionality. Represents the composition of an endmember within a natural system.

see also: SysDraw, EmDraw1, EmDraw2, EmDraw3

EmDraw(x, cx) -> EmDraw1

EmDraw(x, cx, y, cy) -> EmDraw2

EmDraw(x, cx, y, cy, z, cz) -> EmDraw3

The constructor function accepts values as a list of arguments or as keyword assignments, e.g.: EmDraw(x=1, cx=2)

IsoMix.EmDraw1 — Type

EmDraw1 <: EmDraw

1-dimensional EmDraw instance.

Fields

x	Isotopic composition of x

cx	Concentration of x

see also: EmDraw

IsoMix.EmDraw2 — Type

EmDraw2 <: EmDraw

2-dimensional EmDraw instance.

Fields

x	Isotopic composition of x

cx	Concentration of x

y	Isotopic composition of y

cy	Concentration of y

see also: EmDraw

IsoMix.EmDraw3 — Type

EmDraw3 <: EmDraw

3-dimensional EmDraw instance.

Fields

x	Isotopic composition of x

cx	Concentration of x

y	Isotopic composition of y

cy	Concentration of y

z	Isotopic composition of z

cz	Concentration of z

see also: EmDraw

IsoMix.Endmember — Type

Endmember <: Data

Abstract supertype for Endmember_ instances, where the _ indicates dimensionality.

see also: Endmember1, Endmember2, Endmember3

IsoMix.Endmember1 — Type

Endmember1 <: Endmember

1-dimensional Endmember instance.

Fields

x	Isotopic composition of x

cx	Concentration of x

see also: Endmember, Datum

IsoMix.Endmember2 — Type

Endmember2 <: Endmember

2-dimensional Endmember instance.

Fields

x	Isotopic composition of x

cx	Concentration of x

y	Isotopic composition of y

cy	Concentration of y

see also: Endmember, Datum

IsoMix.Endmember3 — Type

Endmember3 <: Endmember

3-dimensional Endmember instance.

Fields

x	Isotopic composition of x

cx	Concentration of x

y	Isotopic composition of y

cy	Concentration of y

z	Isotopic composition of z

cz	Concentration of z

see also: Endmember, Datum

IsoMix.Fraction — Type

Fraction <: IsoMixType

Abstract supertype for Fraction_ instances that contain vectors of fractional combinations of _ components/endmembers.

see also: Fraction2, Fraction3

Fraction(Amin, Amax; n=101)

Returns a Fraction2 with n linearly spaced fractional combinations of components A and B that each sum to 1, given minimum and maximum A fractions of Amin and Amax, respectively. Sincen gives the linear fractions of each component, this constructor returns n-length vectors of each component. Both Amin and Amax must fall within [0,1].

Example

julia> Fraction(0,1,n=3)
Fraction2([0.0, 0.5, 1.0], [1.0, 0.5, 0.0], 3)

Fraction(Amin, Amax, Bmin, Bmax; n=101)

Returns a Fraction3 with linearly spaced fractional combinations of components A, B, and C that each sum to 1, given respective minima and maxima of A (Amin and Amax) and B (Bmin and Bmax). Since n gives the linear fractions of each component, this constructor returns a n(n÷2+1)-length vector for each component. Amin,Amax, Bmin, and Bmax must fall within [0,1].

Example

julia> f=Fraction(0,1,0,1,n=3)
Fraction3([0.0, 0.0, 0.0, 0.5, 0.5, 1.0], [0.0, 0.5, 1.0, 0.0, 0.5, 0.0], [1.0, 0.5, 0.0, 0.5, 0.0, 0.0], 6)

Fraction(dim, n=101)

Given a provided number of components/dimensions dim (2 or 3), returns a corresponding Fraction instance with component concentrations of 0 to 1.

Example

julia> Fraction(2,n=3)
Fraction2([0.0, 0.5, 1.0], [1.0, 0.5, 0.0], 3)

IsoMix.Fraction2 — Type

Fraction2 <: Fraction

2-dimensional Fraction instance.

Fields

A	Endmember/component A

B	Endmember/component B

see also: Fraction

IsoMix.Fraction3 — Type

Fraction3 <: Fraction

3-dimensional Fraction instance.

Fields

A	Endmember/component A

B	Endmember/component B

C	Endmember/component C

see also: Fraction

IsoMix.IsoMixType — Type

IsoMixType

Supertype containing all custom types and structs in the IsoMix package:

Direct subtypes: EmDraw, Endmember, Measurements, Datum, Fraction, Measurements, Model,Prior, SysDraw

IsoMixType
├─ EmDraw
│  ├─ EmDraw1
│  ├─ EmDraw2
│  └─ EmDraw3
├─ Endmember
│  ├─ Endmember1
│  ├─ Endmember2
│  └─ Endmember3
├─ Datum
│  ├─ Constant
│  ├─ Norm
│  ├─ Unf
│  └─ logNorm
├─ Fraction
│  ├─ Fraction2
│  └─ Fraction3
├─ Measurements
├─ Model
│  ├─ Model1
│  ├─ Model2
│  └─ Model3
├─ Prior
│  ├─ Prior2
│  └─ Prior3
└─ SysDraw
   ├─ SysDraw2
   └─ SysDraw3

Generated with

using AbstractTrees,IsoMix; AbstractTrees.children(d) = subtypes(d); print_tree(IsoMixType)

IsoMix.Measurements — Type

Measurements <: Data

Abstract supertype for Measurements_ instances, where _ indicates dimensionality. Represents a dataset of measured (normally distributed) data as vectors of Norm instances. Absent Measurement fields are represented as vectors of Unconstrained and do not affect loglikelihood calculations.

see also: Measurements, Measurements1, Measurements2, Measurements3

Construction

The constructor function Measurements(...) takes nx2 matrices, where n represents the number of measured samples in the dataset, and n must be the same for all provided matrices. Each matrix represents a series of measurements (e.g. isotopic composition or abundance measurement) for a given species. The table below defines all possible fields of a Measurements and which Measurements subtypes contain them (denoted with a ✓):

Fields		Measurements1	Measurements2	Measurements3

x	Isotopic composition of x	✓	✓	✓

cx	Concentration of x	✓	✓	✓

y	Isotopic composition of y		✓	✓

cy	Concentration of y		✓	✓

z	Isotopic composition of z			✓

cz	Concentration of z			✓

Note that each row among the Measurement fields corresponds to a specific sample, so provided matrices must consistently reflect the same sample in each row. Measurements with missing values must be replaced with NaN.

Inputing data into Measurements()

The constructor function accepts a list of Measurements instances:

Without concentration measurements:

Measurements(x, y) -> Measurements2 
Measurements(x, y, z) -> Measurements3

With concentration measurements:

Measurements(x=..., cx=...) -> Measurements1
Measurements(x, cx, y, cy) -> Measurements2
Measurements(x, cx, y, cy, z, cz) -> Measurements3

For all other datasets configurations, declare measurements with keywords:

Measurements(; x, cx, y, cy, z, cz)

...which determines the appropriate subtype and all undeclared fields return vectors of Unconstrained instances.

Examples

julia> Measurements(x = [1 2; 3 4], cx = [5 6; 7 8])
Measurements1(Datum[Norm(1.0, 2.0), Norm(3.0, 4.0)], Datum[Norm(5.0, 6.0), Norm(7.0, 8.0)])

julia> m2d = Measurements([1 2; 3 4], [5 6; 7 NaN]); 

julia> m2d.x
2-element Vector{Datum}:
 Norm(1.0, 2.0)
 Norm(3.0, 4.0)

julia> m2d.cx
2-element Vector{Datum}:
 Unconstrained()
 Unconstrained()

julia> m2d.y
2-element Vector{Datum}:
 Norm(5.0, 6.0)
 Norm(7.0, NaN)

julia> m2d = Measurements([1 2; 3 4], [5 6; 7 8; 9 10])
ERROR: AssertionError: x & y must have the same number of entries (rows). Use NaN for missing values.

IsoMix.Measurements1 — Type

Measurements1 <: Measurements

Measurements instance reflecting isotope and abundance measurements of 1 species: x.

IsoMix.Measurements2 — Type

Measurements2 <: Measurements

Measurements instance reflecting isotope and abundance measurements of 2 species: x, y.

IsoMix.Measurements3 — Type

Measurements3 <: Measurements

Measurements instance reflecting isotope and abundance measurements of 3 species: x, y, z.

IsoMix.Model — Type

Model <: IsoMixType

Abstract supertype for Model_ instances, where the _ indicates dimensionality.

see also: Model1, Model2, Model3

Model(f::Fraction, s::SysDraw)

Generate a Model instance

IsoMix.Model1 — Type

Model1 <: Model

1-dimensional Model instance: a single species tracking isotopic composition and concentration.

Fields

x	Isotopic composition of x

cx	Concentration of x

see also: Model

IsoMix.Model2 — Type

Model2 <: Model

2-dimensional Model instance: two species tracking isotopic composition and concentration.

Fields

x	Isotopic composition of x

cx	Concentration of x

y	Isotopic composition of y

cy	Concentration of y

see also: Model

IsoMix.Model3 — Type

Model3 <: Model

3-dimensional Model instance: three species tracking isotopic composition and concentration.

Fields

x	Isotopic composition of x

cx	Concentration of x

y	Isotopic composition of y

cy	Concentration of y

z	Isotopic composition of z

cz	Concentration of z

see also: Model

IsoMix.Norm — Type

Norm(m, s) <: Datum

Normally distributed Datum with mean m and standard deviation s.

IsoMix.Prior — Type

Prior <: Data

Abstract supertype for Prior_ instances, where the _ indicates dimensionality. Represents a suite of [Endmember] data.

see also: Endmember, Prior2, Prior3

Prior(A::Endmember, B::Endmember)

Returns a Prior2

Prior(A::Endmember, B::Endmember, C::Endmember)

Returns a Prior3

Note: each Endmember must be of the same subtype (i.e. consistent number of components within each endmember of a system).

IsoMix.Prior2 — Type

Prior2 <: Prior

2-dimensional Prior instance.

Fields

A	Endmember/component A

B	Endmember/component B

see also: Prior

IsoMix.Prior3 — Type

Prior3 <: Prior

3-dimensional Prior instance.

Fields

A	Endmember/component A

B	Endmember/component B

C	Endmember/component C

see also: Prior

IsoMix.SysDraw — Type

SysDraw <: Data

Abstract supertype for SysDraw_ instances, where the _ indicates dimensionality. Represents a model system composition charaterized by its EmDraw fields.

see also: EmDraw SysDraw2, SysDraw3

The constructor function accepts values as a list of arguments:

SysDraw(A::EmDraw, B::EmDraw) -> SysDraw2

SysDraw(A::EmDraw, B::EmDraw, C::EmDraw) -> SysDraw3

... or as keyword assignments, e.g.: SysDraw(A=..., B=...)

IsoMix.SysDraw2 — Type

SysDraw2 <: SysDraw

2-dimensional SysDraw instance.

Fields

A	Endmember/component A

B	Endmember/component B

see also: SysDraw

IsoMix.SysDraw3 — Type

SysDraw3 <: SysDraw

3-dimensional SysDraw instance.

Fields

A	Endmember/component A

B	Endmember/component B

C	Endmember/component C

see also: SysDraw

IsoMix.Unconstrained — Type

Unconstrained <: Datum

A Datum instance for an unconstrained variable, functionally similar to Unf(-∞,∞) or an unmeasured Measurements datum. Typically, using this value is inadvisable.

IsoMix.Unf — Type

Unf(a, b) <: Datum

Uniformly distributed Datum with minimum a and maximum b, inclusive. I.e., for any $x$ in $Unf(a, b)$, $x \in [a,b]$.

IsoMix.logNorm — Type

logNorm(lm, ls) <: Datum

Log-normally distributed Datum with log-mean lm and log-space standard deviation ls.

IsoMix.countcomponents — Method

IsoMix.countcomponents(c::EmDraw)

Count the number of elemental dimensions (1-3) of c.

Example

julia> IsoMix.countcomponents(EmDraw(1,2,3,4,5,6))
3

IsoMix.extractcomponents — Method

IsoMix.extractcomponents(c)

Return all values from EmDraw instance c.

see also: IsoMix.extractsystem

IsoMix.extractfields — Method

IsoMix.extractfields(s)

Return all SysDraw field names (e.g. :A, :B) and all EmDraw subfield names (e.g. :x, :cx) as single Symbols defining each, e.g. :Ax, :Bcz.

IsoMix.extractsystem — Method

IsoMix.extractsystem(s)

Returns all EmDraw field values within the SysDraw fields of s.

see also: IsoMix.extractcomponents

IsoMix.fractions — Method

fractions(s<:SysDraw2{EmDraw1}, x)

Calculate the fractional contributions of each the two EmDraws in s, given a composition of x.

fractions(s<:SysDraw3{EmDraw2}, x, y)

Calculate the fractional contributions of each the three EmDraws in s, given compositions of x and y.

IsoMix.initialguess — Method

IsoMix.initialguess(p<:Prior)

Returns a SysDraw instance corresponding to provided Prior instance p, assigning initial guesses of each component composition using the central tendancy of each Datum: mean of Norm, log-mean of logNorm, midpoint of Unf, value of Constant, and an arbitrary value of 0.5 for Unconstrained.

IsoMix.initialjump — Method

IsoMix.initialjump(p<:Prior)

Returns a SysDraw instance corresponding to an initial jumping distribution σ for each component's compositions, given provided Prior instance p. Jump σ are assigned for each Datum as follows, the standar deviation (σ) of Norm, the log-σ of logNorm, ¼ the range of Unf, a value of 0 for Constant, and an arbitrary value of 0.1 for Unconstrained.

IsoMix.jump — Method

IsoMix.jump(s<:SysDraw, j<:SysDraw; rng)

Given a guess of SysDraw values s and corresponding jumping distribution scales in j, randomly perturbs one component of s and returns the new guess as well as a tuple containing the jumped EmDraw/endmember (:A, :B, :C), the jumped component composition (e.g. :x, :y, :cx), and the jump value.

Example

julia> jumpedsystem, t = jump(SysDraw(EmDraw(1,2),EmDraw(3,4)), SysDraw(EmDraw(.1,.5),EmDraw(.2,.4)), rng=IsoMix.Random.Xoshiro(1)); 

julia> jumpedsystem
SysDraw2{EmDraw1}(EmDraw1(1.0, 2.349413341845734), EmDraw1(3.0, 4.0))

julia> t
(:A, :cx, 0.34941334184573425)

IsoMix.loglikelihood — Method

loglikelihood(x::Number, D <: Datum)

Calculate the relative¹ log-likelihood that x is drawn from a distribution D, which may be normal (Norm), lognormal (logNorm), or uniform (Unf). Note that log-space constants are dropped from the calculations for efficiency (metropolis only compares log-ratios)

loglikelihood(c<:EmDraw, d<:Endmember)

Calculate the loglikelihood that the species composition/concentrations in c were drawn from the corresponding distributions in d.

loglikelihood(s<:SysDraw , p<:Prior)

Calculate the loglikelihood that the components in s were drawn from the corresponding Endmember in p.

loglikelihood(model<:Model , measurements<:Measurements)

Calculate the likelihood (ℓ) that the (highest likelihood) model compositions were drawn from the distributions in measurements. Uses Polyester.@batch-based multithreading for faster runtimes.

Note that rather than calculating the ℓ of every model composition relative to the corresponding distribution in 'measurements, the three highest ℓ among all instances in amodelare summed for each measurement. This preventsmixtropolisfrom converging on small parameter spaces that only just fit around themeasurements`, while also limiting aliasing problems.

IsoMix.mix! — Method

mix!(m <: Model, s <: SysDraw, f <: Fraction)

In-place version of mix that overwrites fields in m.

see also: SysDraw, Fraction, Model

IsoMix.mix — Method

mix(s<:SysDraw, f<:Fraction)

Calculate mixing between components of system s given fractional mixtures f, returned in a Model instance.

see also: mix, SysDraw, Fraction, Model

Input configurations:

mix(s::SysDraw2{EmDraw1}, f::Fraction2) -> Model1

Two-endmember, 1 species.

mix(s::SysDraw2{EmDraw2}, f::Fraction2) -> Model2

Two-endmember, 2 species.

mix(s::SysDraw2{EmDraw3}, f::Fraction2) -> Model3

Two-endmember, 3 species.

mix(s::SysDraw3{EmDraw2}, f::Fraction3) -> Model2

Three-endmember, 2 species.

mix(s::SysDraw3{EmDraw3}, f::Fraction3) -> Model3

Three-endmember, 3 species.

IsoMix.nanll — Method

IsoMix.nanll(x,D<:Union{Norm,Unconstrained})

Same as loglikelihood, but returns 0.0 instead of NaN.

IsoMix.reassigncomponents — Method

IsoMix.reassigncomponents(si::Symbol, X, A, B)
IsoMix.reassigncomponents(si::Symbol, X, A, B,C)

Returns a SysDraw containing EmDraws A and B (and C), replacing the field denoted by si with X

Examples

julia> IsoMix.reassigncomponents(:A, EmDraw(2,2), EmDraw(1,1), EmDraw(1,1))

SysDraw2{EmDraw1}(EmDraw1(2.0, 2.0), EmDraw1(1.0, 1.0))

IsoMix.stopwatch — Method

IsoMix.stopwatch(i, n, t)

Convenience function for [porewatermetropolis] that returns a String reporting the progress at step i for total steps n with start time t (in s since the epoch).

IsoMix.top3 — Method

IsoMix.top3(y, x1, x2, x3)

Returns the three highest inputs in decreasing order, as long as x1, x2, and x3 are provided in decreasing order. This is optimized for speed and a very specific use case, so entering x values out of order will result in inaccurate ranking.

IsoMix.update — Function

update(s<:SysDraw, si::Symbol, ci::Symbol, v)

Update the EmDraw field ci of SysDraw field si of s with the value v.

Example

julia> update(SysDraw(EmDraw(1,2),EmDraw(3,4)), :B, :x, 12.)
SysDraw2{EmDraw1}(EmDraw1(1.0, 2.0), EmDraw1(12.0, 4.0))